Gatekeeper api

+++ app/src/main/java/gr/gkortsaridis/gatekeeper/Repositories/VaultRepository.kt
@@ -1,6 +1,7 @@
 package gr.gkortsaridis.gatekeeper.Repositories
 
-import com.google.firebase.auth.FirebaseUser
+import android.annotation.SuppressLint
+import android.util.Log
 import com.google.firebase.firestore.FirebaseFirestore
 import gr.gkortsaridis.gatekeeper.Entities.Vault
 import gr.gkortsaridis.gatekeeper.Entities.VaultColor
@@ -9,68 +10,53 @@ import gr.gkortsaridis.gatekeeper.Interfaces.VaultCreateListener
 import gr.gkortsaridis.gatekeeper.Interfaces.VaultEditListener
 import gr.gkortsaridis.gatekeeper.Interfaces.VaultRetrieveListener
 import gr.gkortsaridis.gatekeeper.Interfaces.VaultSetupListener
+import gr.gkortsaridis.gatekeeper.Utils.GateKeeperAPI
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.plugins.RxJavaPlugins
+import io.reactivex.schedulers.Schedulers
 
+@SuppressLint("CheckResult")
 object VaultRepository {
 
     val allVaults =  Vault("-1", AuthRepository.getUserID(), "All Vaults", VaultColor.White)
 
-    fun setupVaultsForNewUser(user: FirebaseUser, listener: VaultSetupListener) {
+    fun setupVaultsForNewUser(userId: String, listener: VaultSetupListener) {
         val personalVault = Vault(id = "", name = "Personal", account_id = AuthRepository.getUserID(), color = VaultColor.White)
-        retrieveVaultsByAccountID(user.uid, object: VaultRetrieveListener {
-            override fun onVaultsRetrieveSuccess(vaults: ArrayList<Vault>) {
-                if (vaults.size > 0) {
-                    listener.onVaultSetupComplete()
-                }else{
-                    createVault(personalVault, object : VaultCreateListener {
-                        override fun onVaultCreated(vault: Vault) { listener.onVaultSetupComplete() }
-                        override fun onVaultCreateError() { listener.onVaultSetupError() }
-                    })
-                }
-            }
-
-            override fun onVaultsRetrieveError(e: Exception) {
-                createVault(personalVault, object : VaultCreateListener {
-                    override fun onVaultCreated(vault: Vault) { listener.onVaultSetupComplete() }
-                    override fun onVaultCreateError() { listener.onVaultSetupError() }
-                })
-            }
-        })
+        GateKeeperAPI.api.createVault(SecurityRepository.createEncryptedDataRequestBody(personalVault))
+            .subscribeOn(Schedulers.newThread())
+            .observeOn(Schedulers.newThread())
+            .subscribe (
+                {
+                    if (it.errorCode == -1) { listener.onVaultSetupComplete() }
+                    else { listener.onVaultSetupError(it.errorCode, it.errorMsg) }
+                },
+                { listener.onVaultSetupError(it.hashCode(), it.localizedMessage ?: "") }
+            )
     }
 
     fun createVault(vault: Vault, listener: VaultCreateListener) {
-        val db = FirebaseFirestore.getInstance()
 
-        db.collection("vaults")
-            .add(hashMapOf( "account_id" to AuthRepository.getUserID(), "vault" to SecurityRepository.encryptObjectWithUserCredentials(vault) ))
-            .addOnCompleteListener {
-                if (it.isSuccessful) {
-                    vault.id = it.result!!.id
-                    listener.onVaultCreated(vault)
-                }
-                else { listener.onVaultCreateError() }
-            }
-    }
-
-    fun retrieveVaultsByAccountID(accountID: String, retrieveListener: VaultRetrieveListener) {
-
-        val db = FirebaseFirestore.getInstance()
-        db.collection("vaults")
-            .whereEqualTo("account_id",accountID)
-            .get().addOnSuccessListener { result ->
-                val vaultsResult = ArrayList<Vault>()
-                for (document in result) {
-                    val encryptedVault = (document["vault"] ?: "") as String
-                    val decryptedVault = SecurityRepository.decryptStringToObjectWithUserCredentials(encryptedVault, Vault::class.java) as Vault?
-                    if (decryptedVault != null) {
-                        decryptedVault.id = document.id
-                        vaultsResult.add(decryptedVault)
+        GateKeeperAPI.api.createVault(SecurityRepository.createEncryptedDataRequestBody(vault))
+            .subscribeOn(Schedulers.newThread())
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribe (
+                {
+                    if (it.errorCode == -1) {
+                        val encryptedData = it.data
+                        val vault = SecurityRepository.decryptEncryptedDataToObjectWithUserCredentials(encryptedData, Vault::class.java) as Vault?
+                        if (vault != null) {
+                            vault.id = it.data.id.toString()
+                            listener.onVaultCreated(vault)
+                        } else {
+                            listener.onVaultCreateError(-1, "Decryption Error")
+                        }
                     }
-                }
-
-                retrieveListener.onVaultsRetrieveSuccess(vaultsResult)
-            }
-            .addOnFailureListener { exception -> retrieveListener.onVaultsRetrieveError(exception) }
-
+                    else {
+                        listener.onVaultCreateError(it.errorCode, it.errorMsg)
+                    }
+                },
+                { listener.onVaultCreateError(it.hashCode(), it.localizedMessage ?: "") }
+            )
     }
 
     fun getVaultByID(id: String): Vault? {
@@ -109,43 +95,64 @@ object VaultRepository {
         vault.name = newName
         vault.color = color
 
-        val vaulthash = hashMapOf(
-            "vault" to SecurityRepository.encryptObjectWithUserCredentials(vault),
-            "account_id" to AuthRepository.getUserID()
-        )
-
-        val db = FirebaseFirestore.getInstance()
-        db.collection("vaults")
-            .document(vault.id)
-            .set(vaulthash)
-            .addOnCompleteListener {
-                listener.onVaultEdited(vault)
-            }
-
+        GateKeeperAPI.api.updateVault(SecurityRepository.createEncryptedDataRequestBody(vault, vault.id))
+            .subscribeOn(Schedulers.newThread())
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribe (
+                {
+                    if (it.errorCode == -1) {
+                        val encryptedData = it.data
+                        val vault = SecurityRepository.decryptEncryptedDataToObjectWithUserCredentials(encryptedData, Vault::class.java) as Vault?
+                        if (vault != null) {
+                            listener.onVaultEdited(vault)
+                        } else {
+                            listener.onVaultEditError(-1, "Decryption Error")
+                        }
+                    }
+                    else {
+                        listener.onVaultEditError(it.errorCode, it.errorMsg)
+                    }
+                },
+                {
+                    listener.onVaultEditError(it.hashCode(), it.localizedMessage ?: "")
+                }
+            )
     }
 
     fun deleteVault(vault: Vault, listener: VaultEditListener) {
-        val db = FirebaseFirestore.getInstance()
-        db.collection("vaults")
-            .document(vault.id)
-            .delete()
-            .addOnCompleteListener {
-                retrieveVaultsByAccountID(vault.account_id, object: VaultRetrieveListener{
-                    override fun onVaultsRetrieveSuccess(vaults: ArrayList<Vault>) {
-                        GateKeeperApplication.vaults = vaults
+
+        //Delete the vault
+        GateKeeperAPI.api.deleteVault(vaultId = vault.id, body = SecurityRepository.createUsernameHashRequestBody())
+            .subscribeOn(Schedulers.newThread())
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribe (
+                {
+                    if (it.errorCode == -1 && vault.id.toLong() == it.deletedItemID) {
+
+                        //Delete all vault's logins
+                        val vaultLogins = LoginsRepository.filterLoginsByVault(GateKeeperApplication.logins, vault)
+                        for (login in vaultLogins) {
+                            LoginsRepository.deleteLogin(login, null)
+                        }
+
+                        //Delete all vault's cards
+                        val vaultCards = CreditCardRepository.filterCardsByVault(vault)
+                        for (card in vaultCards) {
+                            CreditCardRepository.deleteCreditCard(card, null)
+                        }
+
+                        //Delete all vault's notes
+                        val vaultNotes = NotesRepository.filterNotesByVault(vault)
+                        for (note in vaultNotes) {
+                            NotesRepository.deleteNote(note, null)
+                        }
+
                         listener.onVaultDeleted()
                     }
-
-                    override fun onVaultsRetrieveError(e: Exception) {
-                        listener.onVaultDeleted()
-                    }
-                })
-            }
-
-        val vaultLogins = LoginsRepository.filterLoginsByVault(GateKeeperApplication.logins, vault)
-        for (login in vaultLogins) {
-            LoginsRepository.deleteLogin(login, null)
-        }
+                    else { listener.onVaultDeleteError(it.errorCode, it.errorMsg) }
+                },
+                { listener.onVaultDeleteError(it.hashCode(), it.localizedMessage ?: "") }
+            )
     }
 
 }
